pyproject.toml
[project]
name = "data-engineering-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "airbyte>=0.29.0",
    "dbt-snowflake>=1.10.0",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "snowflake-connector-python>=3.17.2",
    "typer>=0.17.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
]
